+++ testsuite/forge/src/interface/swarm.rs
@@ -3,8 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    interface::system_metrics::SystemMetricsThreshold, AptosPublicInfo, ChainInfo, FullNode,
-    NodeExt, Result, SwarmChaos, Validator, Version,
+    AptosPublicInfo, ChainInfo, FullNode, NodeExt, Result, SwarmChaos, Validator, Version,
 };
 use anyhow::{anyhow, bail};
 use aptos_config::config::NodeConfig;
@@ -84,13 +83,6 @@ pub trait Swarm: Sync {
 
     async fn ensure_no_validator_restart(&self) -> Result<()>;
     async fn ensure_no_fullnode_restart(&self) -> Result<()>;
-
-    async fn ensure_healthy_system_metrics(
-        &mut self,
-        start_time: i64,
-        end_time: i64,
-        threshold: SystemMetricsThreshold,
-    ) -> Result<()>;
 
     // Get prometheus metrics from the swarm
     async fn query_metrics(
